LayoutTests/http/tests/site-isolation/page-lifecycle/unload-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->

<!DOCTYPE html>
<title>Tests that unload is fired in iframe when site isolation is enabled</title>

<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that unload is fired in iframe when site isolation is enabled.");
jsTestIsAsync = true;

const chan = new BroadcastChannel("channel");
chan.addEventListener("message", (e) => {
  caughtEvent = e.data;
  shouldBeEqualToString("caughtEvent.name", "unload");
  finishJSTest();
})

</script>

<iframe src="/site-isolation/page-lifecycle/resources/navigating-away-subpage.html?event=unload"></iframe>
